pyproject.toml
[project]
name = "cs262_design3"
version = "0.1.0"
description = "COMPSCI 262 Design Exercise 3"
readme = "README.md"
python_requires = ">=3.11"
dependencies = [
    "pyyaml",
    "grpcio",
    "grpcio-tools",
]